Rename top level package handlers to subcommands
On conflict, this package is aliased as "api_subcommands".
