+++ subcommands/cat/cat.go
