+++ subcommands/info/errors.go
