+++ subcommands/info/object.go
