+++ subcommands/info/packfile.go
