+++ subcommands/info/snapshot.go
