+++ subcommands/info/state.go
