+++ subcommands/info/vfs.go
